% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin.R
\name{beautifyaddin}
\alias{beautifyaddin}
\title{Beatuify Addin}
\usage{
beautifyaddin()
}
\description{
Addin for beautifyR. It collects the selected markdown table and hands it off
to the `beautifyR` function.
}
\keyword{internal}
