% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautifyComment.R
\name{beautifyComment}
\alias{beautifyComment}
\title{Beautify R comments}
\usage{
beautifyComment(inputstring)
}
\arguments{
\item{inputstring}{Character vector of length 1 containing R code.}
}
\description{
This function beautifys multi-line R comments to fit the 80 column margin 
perfectly without the need of tedious manual formatting.
}
