% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_beautify_comment.R
\name{beautifycommentaddin}
\alias{beautifycommentaddin}
\title{Beatuify Comment Addin}
\usage{
beautifycommentaddin()
}
\description{
Addin for beautifyR. It collects the selected R code and hands it off
to the `beautifyComment` function.
}
\keyword{internal}
