% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautifyR.R
\name{beautifyR}
\alias{beautifyR}
\title{Beautify RMarkdown tables}
\usage{
beautifyR(inputstring)
}
\arguments{
\item{inputstring}{Charactervector of length 1 containing a RMarkdown table.}
}
\description{
This function beautifys RMarkdown tables in terms of columnwidths, alignment,
missing columns and rows. It is called by the beautifyR RStudio addin.
}
