% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addin_beautify_table.R
\name{beautifytableaddin}
\alias{beautifytableaddin}
\title{Beatuify Table Addin}
\usage{
beautifytableaddin()
}
\description{
Addin for beautifyR. It collects the selected markdown table and hands it off
to the `beautifyR` function.
}
\keyword{internal}
