% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAlignment.R
\name{extractAlignment}
\alias{extractAlignment}
\title{Extract Alignment from list of RMarkdown table cells cells}
\usage{
extractAlignment(cells, mc = maxColumns)
}
\arguments{
\item{cells}{list of markdown cells created within `beautifyR`}

\item{mc}{number of columns within the RMarkdown table, derived within
`beautifyR`}
}
\description{
The function extracts the alignment of all cells and returns a character
vector indicating left, right or center alignment for all columns according
to the formatting line (i.e. the 2nd)
}
\keyword{internal}
