% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padCells.R
\name{padCells}
\alias{padCells}
\title{paddCells}
\usage{
padCells(ce, al, mch, mco)
}
\arguments{
\item{ce}{list of table cells as created by `beautifyR`}

\item{al}{alignment vector as created by `extractAlignment`}

\item{mch}{maximum number of characters in each column (defines the column
width)}

\item{mco}{number of columns in the table}
}
\description{
Function to padd cells according to the column width
}
\keyword{internal}
