% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineFormatting.R
\name{refineFormatting}
\alias{refineFormatting}
\title{refineFomatting}
\usage{
refineFormatting(ce, al = align)
}
\arguments{
\item{ce}{list of table cells as created by `beautifyR`}

\item{al}{alignment vector as created by `extractAlignment`}
}
\description{
Function to refine the width of the formatting row (i.e. the 2nd). It
contains information about alignment of the table columns.
}
\keyword{internal}
